refactor: rename the use case to deleteByRecordNumber
