+++ src/record/delete/index.ts
@@ -1,8 +1,13 @@
 import type { RestAPIClientOptions } from "../../kintone/client";
+import type { KintoneRestAPIClient } from "@kintone/rest-api-client";
+import type { RecordNumber } from "./types/field";
+import type { FieldsJson } from "../../kintone/types";
 import { buildRestAPIClient } from "../../kintone/client";
 import { deleteAllRecords } from "./usecases/deleteAll";
-import { deleteRecordsByFile } from "./usecases/delete";
+import { deleteByRecordNumber } from "./usecases/deleteByRecordNumber";
 import { logger } from "../../utils/log";
+import { readFile } from "../../utils/file";
+import { parseRecords } from "./parsers";
 
 export type Options = {
   app: string;
@@ -25,4 +30,52 @@ export const run: (
     // eslint-disable-next-line no-process-exit
     process.exit(1);
   }
+};
+
+const deleteRecordsByFile = async (
+  apiClient: KintoneRestAPIClient,
+  app: string,
+  filePath: string
+): Promise<void> => {
+  const recordNumbers = await getRecordNumbersFromFile(
+    apiClient,
+    app,
+    filePath
+  );
+  if (recordNumbers.length === 0) {
+    logger.warn("The specified CSV file does not have any records.");
+    return;
+  }
+
+  await deleteByRecordNumber(apiClient, app, recordNumbers);
+};
+
+const getRecordNumbersFromFile = async (
+  apiClient: KintoneRestAPIClient,
+  app: string,
+  filePath: string
+): Promise<RecordNumber[]> => {
+  const fieldsJson = await apiClient.app.getFormFields({ app });
+  const recordNumberFieldCode = getRecordNumberFieldCode(fieldsJson.properties);
+  const { content, format } = await readFile(filePath);
+
+  return parseRecords({
+    source: content,
+    format,
+    recordNumberFieldCode,
+  });
+};
+
+const getRecordNumberFieldCode = (
+  properties: FieldsJson["properties"]
+): string => {
+  let recordNumberFieldCode = "";
+  for (const property of Object.values(properties)) {
+    if (property.type === "RECORD_NUMBER") {
+      recordNumberFieldCode = property.code;
+      break;
+    }
+  }
+
+  return recordNumberFieldCode;
 };
